Reorganise transaction handler
