+++ apps/transactions/main.go
@@ -4,6 +4,7 @@ import (
 	"log"
 
 	"github.com/aws/aws-lambda-go/lambda"
+	"github.com/jonDufty/budget/apps/transactions/handler"
 	"github.com/kelseyhightower/envconfig"
 )
 
@@ -15,14 +16,14 @@ func Hello(name string) string {
 func main() {
 	cfg := MustLoadConfig()
 	log.Println(cfg)
-	client := NewTransactionClient(cfg)
+	client := handler.NewTransactionClient(cfg)
 	client.MustPing()
 
 	lambda.Start(client.TransactionHandler)
 }
 
-func MustLoadConfig() Config {
-	var cfg Config
+func MustLoadConfig() handler.Config {
+	var cfg handler.Config
 	err := envconfig.Process("transaction", &cfg)
 	if err != nil {
 		log.Fatal(err.Error())
