+++ libs/database/models/account.go
@@ -1,6 +1,6 @@
 package models
 
 type Account struct {
-	Id   string `meddler:id`
-	Name string `meddler:name`
+	Id   string `meddler:"id"`
+	Name string `meddler:"name"`
 }
