+++ libs/database/models/budgets.go
@@ -1,6 +1,6 @@
 package models
 
 type Budget struct {
-	Category string `meddler:category`
-	Limit    int    `meddler:limit`
+	Category string `meddler:"category"`
+	Limit    int    `meddler:"limit"`
 }
