+++ libs/database/models/merchant.go
@@ -3,9 +3,9 @@ package models
 import "github.com/jonDufty/budget/libs/upbank/client"
 
 type Merchant struct {
-	Name       string `meddler:name`
-	Category   string `meddler:category`
-	UpCategory string `meddler:up_category`
+	Name       string `meddler:"name"`
+	Category   string `meddler:"category"`
+	UpCategory string `meddler:"up_category"`
 }
 
 func NewMerchantFromApi(r client.TransactionResource) *Merchant {
